Added SNMP walk (via both GETBULK and GETNEXT)
The original 'alouca' version has a limited walk implementation. This
version's logic is more robust and is able to successfully walk large
trees across a range of different devices/agents. Flow is based off
existing walk implementations found in net-snmp and Java snmp.

TODO: To make this just as reliable as other implementations, we'll need to
add retry support at the send() layer, and ensure packet request IDs are
matched.