Made INTEGER parsing in parseRawField consistent
Prasing of INTEGER values in parseRawField was incorrect.  It would
returned an unsigned type if it was more than one byte.  This lead to
issues where say a RequestID was < 256 it would return an int, while if it
was > 256 it would return a uint.

The logic in parseRaw is now consistent with decodeValue.