very subtle scoping issue
The err variable was erroneously (re)created in the scope of the loop body. While this didn't affect returning it from the loop, the non-fatal err assignments from the loop to the outer scope variable were now masked, resulting in (nil, nil) function returns, which causes a panic in walk.go:47 when a field in the nil result is accessed.