use a 64k buffer for reading
This commit changes the `receive()` method to use a static 64k buffer.
This is done to alleviate the issue that arises where if we try to read a UDP message that is larger than the buffer, then the message is truncated. 64k is the maximum size of a IP packet (even when fragmented), so that is the size of our buffer.
We then persist the buffer between `receive()` calls by placing it on the `GoSNMP` struct. This is so that the garbage collector is not constantly having to clean them up. In the future, if the golang compiler escape detector is smart enough to figure out that the buffer doesn't escape the function, we could declare the buffer within `receive()`, and let it sit on the stack. But for the moment the escape detector is not smart enough to do this.