eliminated timing issue in TestSendTrap case
I have eliminated the timing issue in the test case by adding a
"listening" flag to the TrapListener struct, along with the associated
sync protection so that it can be waited on.

The new fields are unexported, so they do not affect the API. I can't
think of any other solution considering that TrapListener.Listen is a
blocking call.