Add support for Walk and Get over TCP
RFC 3430 specifies the transport of SNMP over TCP. It can be used to
support more efficient bulk transfers. This is supported in net-snmp and
the snmpd from OpenBSD for example.

This commit adds the field Transport to the GoSNMP struct that allows
the selection of UDP or TCP as transport layer. Connect() uses the
specified method to query the server. The interface of the library
remains the same.

As the connection can be closed at any time by the server, we need to
handle an EOF from the connection gracefully. This change also adds an
example to walk using TCP. It was tested against the snmpd from OpenBSD.