Unmarshal OctetString as a string, not a []byte
Strings are comparable, byte slices are not. This comes at the cost of a
copy (which is probably preferable anyway to avoid memory fragmentation)
and potential breakage of applications which relied on the old
(undocumented) behavior, but with the benefit of both making the
resulting unmarshalled PDUs comparable with other things as well as
mirroring the source data for tests (which use strings).