Band-Aid over global state breakage on trap tests
Too many of the tests make use of the global Default object, which
leads to unpredictable interactions between tests.  This commit makes a
quick sticking plaster over the issue which lets the tests run for
traps, but the problem is deeper; reliance on global state needs to be
sharply reduced or more problems like this will creep in over time.