Initial net-snmp validation testing
running go test with -tags=netsnmp will cause the test to use packets generated by net-snmp as expected values. This requires libsnmpd to be installed and cgo. Adding -rec=true to this will also record the expected packets as files under /testdata. These files can then be used without cgo in normally testing.

An additional pcap generating feature was also added, when -pcap $DIR is added to go test that directory will be filled with the expected and generated packets as pcaps. This allows for debugging using wireshark's parsers.

Signed-off-by: Liam Kelly <liam@commdevices.com>