Allow global password cache to be turned off
The global password cache is undesirable for higher security
applications as a user's cache entry would remain even if the calling
package deleted the corresponding `UsmSecurityParameters`.

Disabling this mechanism comes with a significant performance hit, as
seen by running the added benchmark tests with `NOPWDCACHE=true`.

There is also a middle ground approach, where the calling package can
reset the cache by turning it off then on again. This should clear any
stale entries and still allow the cache to be leveraged.

Signed-off-by: Liam Kelly <liam@commdevices.com>