Prevent panic during TCP EOF (#521)
When performing SNMP over TCP, a panic could have been triggered
due to a nil pointer dereference in marshal.go:send().

The following example could have led to a panic:
- Perform a GET SNMP request over TCP
- Wait for the response in the `waitingResponse` loop
- `(*GoSNMP).receive()` returns io.EOF
- A new tcp connection will be established successfully via
  `(*GoSNMP).netConnect` and `err` will be set to `nil`
- Break out of the `waitingResponse` loop
- `err == nil` and `result == nil`, because latter would
  be initialized only after a successful `(*GoSNMP).receive()`
- Reach the "Success!" point and erroneously `return nil, nil`

This will lead to a nil pointer dereference in the calling
function `send`.

This patch fixes the bug, by actually performing the retry and
continuing the outer retry loop.

The patch also removes the decrementation of the `retries` counter,
as this:

> // EOF on TCP: reconnect and retry. Do not count
> // as retry as socket was broken

is seen a legitimate failed request attempt.

Also, remove the decrementation of `retries` here:
it could potentially lead to an infinite loop, if reconnecting to
the server always succeeds but also always results in an immediate
EOF upon receiving the response.

Also, ensure that `err` is always set if the `retries` counter
exceeds `(*GoSNMP).Retries.

The added test is the work of @TimRots.

Signed-off-by: Stephan Gabert <stepga@nirgendwo.eu>